@@ -1,8 +1,8 @@
 [package]
 name = "forc"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["Fuel Labs <contact@fuel.sh>"]
-edition = "2018"
+edition = "2021"
 homepage = "https://fuel.network/"
 license = "Apache-2.0"
 repository = "https://github.com/FuelLabs/sway"
@@ -12,35 +12,39 @@ description = "Fuel Orchestrator."
 annotate-snippets = { version = "0.9", features = ["color"] }
 ansi_term = "0.12"
 anyhow = "1.0.41"
+clap = { version = "3.1.2", features = ["env", "derive"] }
+clap_complete = "3.1"
 dirs = "3.0.2"
-flate2 = "1.0.20"
-fuel-asm = "0.1" 
+fuel-asm = "0.1"
 fuel-gql-client = { version = "0.3", default-features = false }
 fuel-tx = "0.5"
 fuel-vm = "0.4"
 futures = "0.3"
+git2 = "0.14"
 hex = "0.4.3"
-prettydiff = "0.4.0"
+petgraph = { version = "0.6.0", features = ["serde-1"] }
+prettydiff = "0.5.0"
 reqwest = { version = "0.11.4", default-features = false, features = ["json", "rustls-tls"] }
 semver = "1.0.3"
 serde = {version = "1.0", features = ["derive"]}
 serde_json = "1.0.73"
-structopt = "0.3"
-sway-core = { version = "0.4.0", path = "../sway-core" }
-sway-fmt = { version = "0.4.0", path = "../sway-fmt" }
-sway-server = { version = "0.4.0", path = "../sway-server" }
-sway-utils = { version = "0.4.0", path = "../sway-utils" }
-sway-types = { version = "0.4.0", path = "../sway-types" }
+sway-core = { version = "0.5.0", path = "../sway-core" }
+sway-fmt = { version = "0.5.0", path = "../sway-fmt" }
+sway-server = { version = "0.5.0", path = "../sway-server" }
+sway-utils = { version = "0.5.0", path = "../sway-utils" }
+sway-types = { version = "0.5.0", path = "../sway-types" }
 taplo = "0.7"
 tar = "0.4.35"
 term-table = "1.3"
 termcolor = "1.1"
 tokio = { version = "1.8.0", features = ["macros", "rt-multi-thread", "process"] }
 toml = "0.5"
+unicode-xid = "0.2.2"
 ureq = "2.4"
+url = "2"
+uwuify = { version = "^0.2", optional = true }
 warp = "0.3"
 whoami = "1.1"
-uwuify = { version = "^0.2", optional = true }
 
 [features]
 default = []
