@@ -1,17 +1,18 @@
 [package]
 name = "sway-core"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["Fuel Labs <contact@fuel.sh>"]
-edition = "2018"
+edition = "2021"
 homepage = "https://fuel.network/"
 license = "Apache-2.0"
 repository = "https://github.com/FuelLabs/sway"
 description = "Sway core language."
 
 [features]
-selector-debug = ["structopt", "hex"]
+selector-debug = ["clap", "hex"]
 
 [dependencies]
+clap = { version = "3.1.2", features = ["derive"], optional = true }
 derivative = "2.2.0"
 dirs = "3.0"
 either = "1.6"
@@ -23,15 +24,16 @@ lazy_static = "1.4"
 nanoid = "0.4"
 pest = { version = "3.0.4", package = "fuel-pest" }
 pest_derive = { version = "3.0.4", package = "fuel-pest_derive" }
-petgraph = "0.5"
+petgraph = "0.6"
 prettydiff = "0.5"
+regex = "1"
 serde = { version = "1.0", features = ["derive"] }
 sha2 = "0.9"
 smallvec = "1.7"
 structopt = { version = "0.3", default-features = false, optional = true }
-sway-ir = { version = "0.4.0", path = "../sway-ir" }
-sway-types = { version = "0.4.0", path = "../sway-types" }
-sway-utils = { version = "0.4.0", path = "../sway-utils" }
+sway-ir = { version = "0.5.0", path = "../sway-ir" }
+sway-types = { version = "0.5.0", path = "../sway-types" }
+sway-utils = { version = "0.5.0", path = "../sway-utils" }
 generational-arena = "0.2"
 bigint = "4.4.3"
 thiserror = "1.0"
