@@ -1,8 +1,8 @@
 [package]
 name = "sway-fmt"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["Fuel Labs <contact@fuel.sh>"]
-edition = "2018"
+edition = "2021"
 homepage = "https://fuel.network/"
 license = "Apache-2.0"
 repository = "https://github.com/FuelLabs/sway"
@@ -10,6 +10,6 @@ description = "Sway sway-fmt."
 
 [dependencies]
 ropey = "1.2"
-sway-core = { version = "0.4.0", path = "../sway-core" }
-sway-types = { version = "0.4.0", path = "../sway-types" }
+sway-core = { version = "0.5.0", path = "../sway-core" }
+sway-types = { version = "0.5.0", path = "../sway-types" }
 
