@@ -1,8 +1,8 @@
 [package]
 name = "sway-server"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["Fuel Labs <contact@fuel.sh>"]
-edition = "2018"
+edition = "2021"
 homepage = "https://fuel.network/"
 license = "Apache-2.0"
 repository = "https://github.com/FuelLabs/sway"
@@ -13,8 +13,8 @@ dashmap = "4.0.2"
 lspower = "1.0.0"
 ropey = "1.2"
 serde_json = "1.0.60"
-sway-core = { version = "0.4.0", path = "../sway-core" }
-sway-fmt = { version = "0.4.0", path = "../sway-fmt" }
-sway-types = { version = "0.4.0", path = "../sway-types" }
-sway-utils = { version = "0.4.0", path = "../sway-utils" }
+sway-core = { version = "0.5.0", path = "../sway-core" }
+sway-fmt = { version = "0.5.0", path = "../sway-fmt" }
+sway-types = { version = "0.5.0", path = "../sway-types" }
+sway-utils = { version = "0.5.0", path = "../sway-utils" }
 tokio = { version = "1.3", features = ["io-std", "io-util", "macros", "net", "rt-multi-thread", "sync", "time"] }
