@@ -2,11 +2,12 @@
 name = "test"
 version = "0.0.0"
 authors = ["Fuel Labs <contact@fuel.sh>"]
-edition = "2018"
+edition = "2021"
 publish = false
 
 [dependencies]
-forc = { version = "0.4.0", path = "../forc", features = ["test"], default-features = false }
+anyhow = "1.0.41"
+forc = { version = "0.5.0", path = "../forc", features = ["test"], default-features = false }
 fuel-asm = "0.1"
 fuel-tx = "0.5"
 fuel-vm = { version = "0.4", features = ["random"] }
