@@ -1,24 +1,25 @@
 use crate::ops::forc_explorer;
-use structopt::StructOpt;
+use anyhow::{bail, Result};
+use clap::Parser;
 
 /// Run the network explorer.
-#[derive(Debug, StructOpt)]
+#[derive(Debug, Parser)]
 pub struct Command {
     /// The port number
-    #[structopt(short = "p", long = "port", default_value = "3030")]
+    #[clap(short = 'p', long = "port", default_value = "3030")]
     pub port: String,
-    #[structopt(subcommand)] // Note that we mark a field as a subcommand
+    #[clap(subcommand)] // Note that we mark a field as a subcommand
     pub clean: Option<CleanCommand>,
 }
 
-#[derive(Debug, StructOpt)]
+#[derive(Debug, Parser)]
 pub enum CleanCommand {
     Clean,
 }
 
-pub(crate) async fn exec(_command: Command) -> Result<(), String> {
+pub(crate) async fn exec(_command: Command) -> Result<()> {
     match forc_explorer::exec(_command).await {
-        Err(e) => Err(e.to_string()),
+        Err(e) => bail!(e),
         _ => Ok(()),
     }
 }
