@@ -1,8 +1,11 @@
 use crate::utils::defaults;
+use anyhow::Result;
 use std::fs;
 use sway_utils::constants;
 
-pub(crate) fn init_new_project(project_name: String) -> Result<(), Box<dyn std::error::Error>> {
+pub(crate) fn init_new_project(project_name: String) -> Result<()> {
+    let neat_name: String = project_name.split('/').last().unwrap().to_string();
+
     // Make a new directory for the project
     fs::create_dir_all(format!("{}/src", project_name))?;
 
@@ -12,13 +15,13 @@ pub(crate) fn init_new_project(project_name: String) -> Result<(), Box<dyn std::
     // Insert default manifest file
     fs::write(
         format!("{}/{}", project_name, constants::MANIFEST_FILE_NAME),
-        defaults::default_manifest(&project_name),
+        defaults::default_manifest(&neat_name),
     )?;
 
     // Insert default test manifest file
     fs::write(
         format!("{}/{}", project_name, constants::TEST_MANIFEST_FILE_NAME),
-        defaults::default_tests_manifest(&project_name),
+        defaults::default_tests_manifest(&neat_name),
     )?;
 
     // Insert default main function
@@ -31,6 +34,12 @@ pub(crate) fn init_new_project(project_name: String) -> Result<(), Box<dyn std::
     fs::write(
         format!("{}/tests/harness.rs", project_name),
         defaults::default_test_program(),
+    )?;
+
+    // Ignore default `out` and `target` directories created by forc and cargo.
+    fs::write(
+        format!("{}/.gitignore", project_name),
+        defaults::default_gitignore(),
     )?;
 
     Ok(())
