@@ -1,6 +1,6 @@
 pub mod forc_abi_json;
 pub mod forc_build;
-pub mod forc_dep_check;
+pub mod forc_clean;
 pub mod forc_deploy;
 pub mod forc_explorer;
 pub mod forc_fmt;
