@@ -1,8 +1,10 @@
 use super::manifest::Manifest;
+use crate::utils::restricted_names;
 use annotate_snippets::{
     display_list::{DisplayList, FormatOptions},
     snippet::{Annotation, AnnotationType, Slice, Snippet, SourceAnnotation},
 };
+use anyhow::{anyhow, bail, Result};
 use std::ffi::OsStr;
 use std::io::{self, Write};
 use std::path::{Path, PathBuf};
@@ -11,6 +13,8 @@ use std::sync::Arc;
 use sway_core::{error::LineCol, CompileError, CompileWarning, TreeType};
 use sway_utils::constants;
 use termcolor::{self, Color as TermColor, ColorChoice, ColorSpec, StandardStream, WriteColor};
+
+pub const DEFAULT_OUTPUT_DIRECTORY: &str = "out";
 
 pub fn is_sway_file(file: &Path) -> bool {
     let res = file.extension();
@@ -24,17 +28,21 @@ pub fn find_main_path(manifest_dir: &Path, manifest: &Manifest) -> PathBuf {
     code_dir
 }
 
-pub fn find_file_name<'sc>(manifest_dir: &Path, main_path: &'sc Path) -> Result<&'sc Path, String> {
+pub fn find_file_name<'sc>(manifest_dir: &Path, main_path: &'sc Path) -> Result<&'sc Path> {
     let mut file_path = manifest_dir.to_path_buf();
     file_path.pop();
     let file_name = match main_path.strip_prefix(file_path.clone()) {
         Ok(o) => o,
-        Err(err) => return Err(err.to_string()),
+        Err(err) => bail!(err),
     };
     Ok(file_name)
 }
 
-pub fn read_manifest(manifest_dir: &Path) -> Result<Manifest, String> {
+pub fn lock_path(manifest_dir: &Path) -> PathBuf {
+    manifest_dir.join(constants::LOCK_FILE_NAME)
+}
+
+pub fn read_manifest(manifest_dir: &Path) -> Result<Manifest> {
     let manifest_path = {
         let mut man = PathBuf::from(manifest_dir);
         man.push(constants::MANIFEST_FILE_NAME);
@@ -44,19 +52,72 @@ pub fn read_manifest(manifest_dir: &Path) -> Result<Manifest, String> {
     let manifest = match std::fs::read_to_string(manifest_path) {
         Ok(o) => o,
         Err(e) => {
-            return Err(format!(
-                "failed to read manifest at {:?}: {}",
-                manifest_path_str, e
-            ))
+            bail!("failed to read manifest at {:?}: {}", manifest_path_str, e)
         }
     };
-    match toml::from_str(&manifest) {
+
+    let manifest = match toml::from_str(&manifest) {
         Ok(o) => Ok(o),
-        Err(e) => Err(format!("Error parsing manifest: {}.", e)),
-    }
+        Err(e) => Err(anyhow!("Error parsing manifest: {}.", e)),
+    }?;
+
+    validate_manifest(manifest)
 }
 
-pub fn get_main_file(manifest_of_dep: &Manifest, manifest_dir: &Path) -> Result<Arc<str>, String> {
+// Using (https://github.com/rust-lang/cargo/blob/489b66f2e458404a10d7824194d3ded94bc1f4e4/src/cargo/util/toml/mod.rs +
+// https://github.com/rust-lang/cargo/blob/489b66f2e458404a10d7824194d3ded94bc1f4e4/src/cargo/ops/cargo_new.rs) for reference
+
+fn validate_name(name: &str, use_case: &str) -> Result<()> {
+    // if true returns formatted error
+    restricted_names::contains_invalid_char(name, use_case)?;
+
+    if restricted_names::is_keyword(name) {
+        bail!("the name `{name}` cannot be used as a package name, it is a Sway keyword");
+    }
+    if restricted_names::is_conflicting_artifact_name(name) {
+        bail!(
+            "the name `{name}` cannot be used as a package name, \
+            it conflicts with Forc's build directory names"
+        );
+    }
+    if name == "test" {
+        bail!(
+            "the name `test` cannot be used as a package name, \
+            it conflicts with Sway's built-in test library"
+        );
+    }
+    if restricted_names::is_conflicting_suffix(name) {
+        bail!(
+            "the name `{name}` is part of Sway's standard library\n\
+            It is recommended to use a different name to avoid problems."
+        );
+    }
+    if restricted_names::is_windows_reserved(name) {
+        if cfg!(windows) {
+            bail!("cannot use name `{name}`, it is a reserved Windows filename");
+        } else {
+            bail!(
+                "the name `{name}` is a reserved Windows filename\n\
+                This package will not work on Windows platforms."
+            );
+        }
+    }
+    if restricted_names::is_non_ascii_name(name) {
+        bail!("the name `{name}` contains non-ASCII characters which are unsupported");
+    }
+    Ok(())
+}
+
+fn validate_manifest(manifest: Manifest) -> Result<Manifest> {
+    validate_name(&manifest.project.name, "package name")?;
+    if let Some(ref org) = manifest.project.organization {
+        validate_name(org, "organization name")?;
+    }
+
+    Ok(manifest)
+}
+
+pub fn get_main_file(manifest_of_dep: &Manifest, manifest_dir: &Path) -> Result<Arc<str>> {
     let main_path = {
         let mut code_dir = PathBuf::from(manifest_dir);
         code_dir.push(constants::SRC_DIR);
@@ -65,18 +126,34 @@ pub fn get_main_file(manifest_of_dep: &Manifest, manifest_dir: &Path) -> Result<
     };
 
     // some hackery to get around lifetimes for now, until the AST returns a non-lifetime-bound AST
-    let main_file = std::fs::read_to_string(&main_path).map_err(|e| e.to_string())?;
+    let main_file = std::fs::read_to_string(&main_path).map_err(|e| e)?;
     let main_file = Arc::from(main_file);
     Ok(main_file)
+}
+
+pub fn default_output_directory(manifest_dir: &Path) -> PathBuf {
+    manifest_dir.join(DEFAULT_OUTPUT_DIRECTORY)
+}
+
+/// Returns the user's `.forc` directory, `$HOME/.forc` by default.
+pub fn user_forc_directory() -> PathBuf {
+    dirs::home_dir()
+        .expect("unable to find the user home directory")
+        .join(constants::USER_FORC_DIRECTORY)
+}
+
+/// The location at which `forc` will checkout git repositories.
+pub fn git_checkouts_directory() -> PathBuf {
+    user_forc_directory().join("git").join("checkouts")
 }
 
 pub fn print_on_success(
     silent_mode: bool,
     proj_name: &str,
-    warnings: Vec<CompileWarning>,
-    tree_type: TreeType,
+    warnings: &[CompileWarning],
+    tree_type: &TreeType,
 ) {
-    let type_str = match tree_type {
+    let type_str = match &tree_type {
         TreeType::Script {} => "script",
         TreeType::Contract {} => "contract",
         TreeType::Predicate {} => "predicate",
@@ -104,7 +181,7 @@ pub fn print_on_success(
     }
 }
 
-pub fn print_on_success_library(silent_mode: bool, proj_name: &str, warnings: Vec<CompileWarning>) {
+pub fn print_on_success_library(silent_mode: bool, proj_name: &str, warnings: &[CompileWarning]) {
     if !silent_mode {
         warnings.iter().for_each(format_warning);
     }
@@ -125,16 +202,12 @@ pub fn print_on_success_library(silent_mode: bool, proj_name: &str, warnings: Ve
     }
 }
 
-pub fn print_on_failure(
-    silent_mode: bool,
-    warnings: Vec<CompileWarning>,
-    errors: Vec<CompileError>,
-) {
+pub fn print_on_failure(silent_mode: bool, warnings: &[CompileWarning], errors: &[CompileError]) {
     let e_len = errors.len();
 
     if !silent_mode {
         warnings.iter().for_each(format_warning);
-        errors.into_iter().for_each(|error| format_err(&error));
+        errors.iter().for_each(format_err);
     }
 
     println_red_err(&format!(
@@ -143,6 +216,33 @@ pub fn print_on_failure(
         if e_len > 1 { "errors" } else { "error" }
     ))
     .unwrap();
+}
+
+pub(crate) fn print_lock_diff(proj_name: &str, diff: &crate::lock::Diff) {
+    print_removed_pkgs(proj_name, diff.removed.iter().cloned());
+    print_added_pkgs(proj_name, diff.added.iter().cloned());
+}
+
+pub(crate) fn print_removed_pkgs<'a, I>(proj_name: &str, removed: I)
+where
+    I: IntoIterator<Item = &'a crate::lock::PkgLock>,
+{
+    for pkg in removed {
+        if pkg.name != proj_name {
+            let _ = println_red(&format!("  Removing {}", pkg.unique_string()));
+        }
+    }
+}
+
+pub(crate) fn print_added_pkgs<'a, I>(proj_name: &str, removed: I)
+where
+    I: IntoIterator<Item = &'a crate::lock::PkgLock>,
+{
+    for pkg in removed {
+        if pkg.name != proj_name {
+            let _ = println_green(&format!("    Adding {}", pkg.unique_string()));
+        }
+    }
 }
 
 pub fn println_red(txt: &str) -> io::Result<()> {
@@ -299,6 +399,7 @@ fn construct_window<'a>(
     const NUM_LINES_BUFFER: usize = 2;
 
     let total_lines_in_input = input.chars().filter(|x| *x == '\n').count();
+    debug_assert!(end.line >= start.line);
     let total_lines_of_highlight = end.line - start.line;
     debug_assert!(total_lines_in_input > total_lines_of_highlight);
 
@@ -311,7 +412,7 @@ fn construct_window<'a>(
             current_line += 1
         }
 
-        if current_line >= start.line - NUM_LINES_BUFFER && calculated_start_ix.is_none() {
+        if current_line + NUM_LINES_BUFFER >= start.line && calculated_start_ix.is_none() {
             calculated_start_ix = Some(ix);
             lines_to_start_of_snippet = current_line;
         }
@@ -327,8 +428,8 @@ fn construct_window<'a>(
     let calculated_start_ix = calculated_start_ix.unwrap_or(0);
     let calculated_end_ix = calculated_end_ix.unwrap_or(input.len());
 
-    *start_ix -= calculated_start_ix;
-    *end_ix -= calculated_start_ix;
+    *start_ix -= std::cmp::min(calculated_start_ix, *start_ix);
+    *end_ix -= std::cmp::min(calculated_start_ix, *end_ix);
     start.line = lines_to_start_of_snippet;
     &input[calculated_start_ix..calculated_end_ix]
 }
