@@ -4,8 +4,8 @@ use crate::{
     parse_tree::{CallPath, Visibility},
     semantic_analysis::{
         ast_node::{
-            TypedAbiDeclaration, TypedCodeBlock, TypedConstantDeclaration, TypedDeclaration,
-            TypedEnumDeclaration, TypedExpression, TypedExpressionVariant,
+            SizeOfVariant, TypedAbiDeclaration, TypedCodeBlock, TypedConstantDeclaration,
+            TypedDeclaration, TypedEnumDeclaration, TypedExpression, TypedExpressionVariant,
             TypedFunctionDeclaration, TypedReassignment, TypedReturnStatement,
             TypedStructDeclaration, TypedStructExpressionField, TypedTraitDeclaration,
             TypedVariableDeclaration, TypedWhileLoop,
@@ -512,8 +512,9 @@ fn connect_trait_declaration(
 ) {
     graph.namespace.add_trait(
         CallPath {
-            suffix: decl.name.clone(),
             prefixes: vec![],
+            suffix: decl.name.clone(),
+            is_absolute: false,
         },
         entry_node,
     );
@@ -527,8 +528,9 @@ fn connect_abi_declaration(
 ) {
     graph.namespace.add_trait(
         CallPath {
-            suffix: decl.name.clone(),
             prefixes: vec![],
+            suffix: decl.name.clone(),
+            is_absolute: false,
         },
         entry_node,
     );
@@ -986,6 +988,21 @@ fn connect_expression(
                 Ok(vec![this_ix])
             }
             None => Ok(leaves.to_vec()),
+        },
+        SizeOf { variant } => match variant {
+            SizeOfVariant::Type(_) => Ok(vec![]),
+            SizeOfVariant::Val(exp) => {
+                let exp = connect_expression(
+                    &(*exp).expression,
+                    graph,
+                    leaves,
+                    exit_node,
+                    "size_of",
+                    tree_type,
+                    exp.span.clone(),
+                )?;
+                Ok(exp)
+            }
         },
         a => {
             println!("Unimplemented: {:?}", a);
