@@ -9,6 +9,9 @@ use pest::iterators::Pair;
 pub struct CallPath {
     pub prefixes: Vec<Ident>,
     pub suffix: Ident,
+    // If `is_absolute` is true, then this call path is an absolute path from
+    // the project root namespace. If not, then it is relative to the current namespace.
+    pub(crate) is_absolute: bool,
 }
 
 impl std::convert::From<Ident> for CallPath {
@@ -16,6 +19,7 @@ impl std::convert::From<Ident> for CallPath {
         CallPath {
             prefixes: vec![],
             suffix: other,
+            is_absolute: false,
         }
     }
 }
@@ -66,10 +70,31 @@ impl CallPath {
         pair: Pair<Rule>,
         config: Option<&BuildConfig>,
     ) -> CompileResult<CallPath> {
+        assert!(pair.as_rule() == Rule::call_path || pair.as_rule() == Rule::call_path_);
         let mut warnings = vec![];
         let mut errors = vec![];
+        let span = Span {
+            span: pair.as_span(),
+            path: config.map(|c| c.path()),
+        };
+        if !(pair.as_rule() == Rule::call_path || pair.as_rule() == Rule::call_path_) {
+            errors.push(CompileError::ParseError {
+                span,
+                err: "expected call path here".to_string(),
+            });
+            return err(warnings, errors);
+        }
         let mut pairs_buf = vec![];
-        for pair in pair.into_inner() {
+        let stmt = pair.into_inner().next().unwrap();
+        let is_absolute = stmt.as_rule() == Rule::absolute_call_path
+            || stmt.as_rule() == Rule::absolute_call_path_;
+        let stmt = stmt.into_inner();
+        let it = if is_absolute {
+            stmt.skip(1)
+        } else {
+            stmt.skip(0)
+        };
+        for pair in it {
             if pair.as_rule() != Rule::path_separator {
                 pairs_buf.push(check!(
                     ident::parse_from_pair(pair, config),
@@ -83,7 +108,24 @@ impl CallPath {
         let suffix = pairs_buf.pop().unwrap();
         let prefixes = pairs_buf;
 
-        // TODO eventually we want to be able to call methods with colon-delineated syntax
-        ok(CallPath { prefixes, suffix }, warnings, errors)
+        ok(
+            CallPath {
+                prefixes,
+                suffix,
+                is_absolute,
+            },
+            warnings,
+            errors,
+        )
+    }
+
+    pub(crate) fn friendly_name(&self) -> String {
+        let mut buf = String::new();
+        for prefix in self.prefixes.iter() {
+            buf.push_str(prefix.as_str());
+            buf.push_str("::");
+        }
+        buf.push_str(self.suffix.as_str());
+        buf
     }
 }
