@@ -51,7 +51,6 @@ impl ImplTrait {
         let impl_keyword = iter.next().unwrap();
         assert_eq!(impl_keyword.as_str(), "impl");
         let trait_name = iter.next().unwrap();
-        assert_eq!(trait_name.as_rule(), Rule::trait_name);
         let trait_name = check!(
             CallPath::parse_from_pair(trait_name, config),
             return err(warnings, errors),
@@ -77,11 +76,14 @@ impl ImplTrait {
             errors
         );
 
-        let where_clause_pair = if iter.peek().unwrap().as_rule() == Rule::trait_bounds {
-            iter.next()
-        } else {
-            None
+        let where_clause_pair = match iter.peek() {
+            Some(r) => match r.as_rule() {
+                Rule::trait_bounds => iter.next(),
+                _ => None,
+            },
+            None => None,
         };
+
         let type_arguments_span = match type_params_pair {
             Some(ref x) => Span {
                 span: x.as_span(),
