@@ -6,11 +6,6 @@ use crate::{
     semantic_analysis::ast_node::*,
     type_engine::{insert_type, IntegerBits},
 };
-use sway_types::join_spans;
-
-use either::Either;
-use std::cmp::Ordering;
-use std::collections::{HashMap, HashSet};
 
 mod method_application;
 
@@ -46,7 +41,6 @@ impl TypedExpression {
             self_type,
             build_config,
             dead_code_graph,
-            dependency_graph,
             opts,
             ..
         } = arguments;
@@ -72,7 +66,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 },
@@ -88,7 +81,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 },
@@ -104,7 +96,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             // TODO if _condition_ is constant, evaluate it and compile this to an
@@ -123,7 +114,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     help_text: Default::default(),
                     opts,
@@ -138,7 +128,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::StructExpression {
@@ -154,7 +143,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::SubfieldExpression {
@@ -170,7 +158,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::MethodApplication {
@@ -186,7 +173,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::Tuple { fields, span } => Self::type_check_tuple(
@@ -198,7 +184,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::TupleIndex {
@@ -216,7 +201,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::DelineatedPath {
@@ -234,7 +218,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::AbiCast {
@@ -250,7 +233,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::Array { contents, span } => Self::type_check_array(
@@ -261,7 +243,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             ),
             Expression::ArrayIndex {
@@ -276,7 +257,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     opts,
                     return_type_annotation: insert_type(TypeInfo::Unknown),
                     mode: Default::default(),
@@ -293,7 +273,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     opts,
                 )
             }
@@ -305,13 +284,46 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
+                    opts,
+                    return_type_annotation: insert_type(TypeInfo::Unknown),
+                    mode: Default::default(),
+                    help_text: Default::default(),
+                },
+            ),
+            Expression::SizeOfVal { exp, span } => Self::type_check_size_of_val(
+                TypeCheckArguments {
+                    checkee: *exp,
+                    namespace,
+                    crate_namespace,
+                    self_type,
+                    build_config,
+                    dead_code_graph,
                     opts,
                     return_type_annotation: insert_type(TypeInfo::Unknown),
                     mode: Default::default(),
                     help_text: Default::default(),
                 },
                 &expr_span,
+                span,
+            ),
+            Expression::SizeOfType {
+                type_name,
+                type_span,
+                span,
+            } => Self::type_check_size_of_type(
+                TypeCheckArguments {
+                    checkee: (type_name, type_span),
+                    namespace,
+                    crate_namespace,
+                    self_type,
+                    build_config,
+                    dead_code_graph,
+                    opts,
+                    return_type_annotation: insert_type(TypeInfo::Unknown),
+                    mode: Default::default(),
+                    help_text: Default::default(),
+                },
+                span,
             ),
             a => {
                 let errors = vec![CompileError::Unimplemented(
@@ -398,7 +410,6 @@ impl TypedExpression {
             Literal::Numeric(_) => TypeInfo::Numeric,
             Literal::U8(_) => TypeInfo::UnsignedInteger(IntegerBits::Eight),
             Literal::U16(_) => TypeInfo::UnsignedInteger(IntegerBits::Sixteen),
-
             Literal::U32(_) => TypeInfo::UnsignedInteger(IntegerBits::ThirtyTwo),
             Literal::U64(_) => TypeInfo::UnsignedInteger(IntegerBits::SixtyFour),
             Literal::Boolean(_) => TypeInfo::Boolean,
@@ -481,7 +492,6 @@ impl TypedExpression {
             self_type,
             build_config,
             dead_code_graph,
-            dependency_graph,
             opts,
             ..
         } = arguments;
@@ -493,124 +503,38 @@ impl TypedExpression {
             warnings,
             errors
         );
-        let TypedFunctionDeclaration {
-            parameters,
-            return_type,
-            body,
-            span,
-            purity,
-            ..
-        } = if let TypedDeclaration::FunctionDeclaration(decl) = function_declaration {
-            // if this is a generic function, monomorphize its internal types and insert the resulting
-            // declaration into the namespace. Then, use that instead.
-            if decl.type_parameters.is_empty() {
-                decl
+        let typed_function_decl =
+            if let TypedDeclaration::FunctionDeclaration(decl) = function_declaration {
+                // if this is a generic function, monomorphize its internal types and insert the resulting
+                // declaration into the namespace. Then, use that instead.
+                if decl.type_parameters.is_empty() {
+                    decl
+                } else {
+                    check!(
+                        decl.monomorphize(type_arguments, self_type),
+                        return err(warnings, errors),
+                        warnings,
+                        errors
+                    )
+                }
             } else {
-                check!(
-                    decl.monomorphize(type_arguments, self_type),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                )
-            }
-        } else {
-            errors.push(CompileError::NotAFunction {
-                name: name.span().as_str().to_string(),
-                span: name.span(),
-                what_it_is: function_declaration.friendly_name(),
-            });
-            return err(warnings, errors);
-        };
-
-        if opts.purity != purity {
-            errors.push(CompileError::PureCalledImpure { span: name.span() });
-        }
-
-        match arguments.len().cmp(&parameters.len()) {
-            Ordering::Greater => {
-                let arguments_span = arguments.iter().fold(
-                    arguments
-                        .get(0)
-                        .map(|x| x.span())
-                        .unwrap_or_else(|| name.span()),
-                    |acc, arg| join_spans(acc, arg.span()),
-                );
-                errors.push(CompileError::TooManyArgumentsForFunction {
-                    span: arguments_span,
-                    method_name: name.suffix.clone(),
-                    expected: parameters.len(),
-                    received: arguments.len(),
+                errors.push(CompileError::NotAFunction {
+                    name: name.span().as_str().to_string(),
+                    span: name.span(),
+                    what_it_is: function_declaration.friendly_name(),
                 });
-            }
-            Ordering::Less => {
-                let arguments_span = arguments.iter().fold(
-                    arguments
-                        .get(0)
-                        .map(|x| x.span())
-                        .unwrap_or_else(|| name.span()),
-                    |acc, arg| join_spans(acc, arg.span()),
-                );
-                errors.push(CompileError::TooFewArgumentsForFunction {
-                    span: arguments_span,
-                    method_name: name.suffix.clone(),
-                    expected: parameters.len(),
-                    received: arguments.len(),
-                });
-            }
-            Ordering::Equal => {}
-        }
-        // type check arguments in function application vs arguments in function
-        // declaration. Use parameter type annotations as annotations for the
-        // arguments
-        //
-        let typed_call_arguments = arguments
-            .into_iter()
-            .zip(parameters.iter())
-            .map(|(arg, param)| {
-                (
-                    param.name.clone(),
-                    TypedExpression::type_check(TypeCheckArguments {
-                        checkee: arg.clone(),
-                        namespace,
-                        crate_namespace,
-                        return_type_annotation: param.r#type,
-                        help_text:
-                            "The argument that has been provided to this function's type does \
-                            not match the declared type of the parameter in the function \
-                            declaration.",
-                        self_type,
-                        build_config,
-                        dead_code_graph,
-                        dependency_graph,
-                        mode: Mode::NonAbi,
-                        opts,
-                    })
-                    .unwrap_or_else(&mut warnings, &mut errors, || {
-                        error_recovery_expr(arg.span())
-                    }),
-                )
-            })
-            .collect();
-
-        ok(
-            TypedExpression {
-                return_type,
-                // now check the function call return type
-                // FEATURE this IsConstant can be true if the function itself is
-                // constant-able const functions would be an
-                // advanced feature and are not supported right
-                // now
-                is_constant: IsConstant::No,
-                expression: TypedExpressionVariant::FunctionApplication {
-                    arguments: typed_call_arguments,
-                    name,
-                    function_body: body,
-                    selector: None, // regular functions cannot be in a contract call; only methods
-                },
-                span,
-            },
-            warnings,
-            errors,
+                return err(warnings, errors);
+            };
+        instantiate_function_application(
+            typed_function_decl,
+            name,
+            arguments,
+            namespace,
+            crate_namespace,
+            self_type,
+            build_config,
+            dead_code_graph,
+            opts,
         )
     }
 
@@ -625,7 +549,6 @@ impl TypedExpression {
             self_type,
             build_config,
             dead_code_graph,
-            dependency_graph,
             return_type_annotation,
             opts,
             ..
@@ -645,7 +568,6 @@ impl TypedExpression {
                 return_type_annotation,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
             }),
             error_recovery_expr(lhs.span()),
             warnings,
@@ -662,7 +584,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -697,7 +618,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
@@ -712,7 +632,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -766,7 +685,6 @@ impl TypedExpression {
             self_type,
             build_config,
             dead_code_graph,
-            dependency_graph,
             opts,
             ..
         } = arguments;
@@ -782,7 +700,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -800,7 +717,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -819,7 +735,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 }),
@@ -871,7 +786,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
@@ -906,7 +820,6 @@ impl TypedExpression {
                                 self_type,
                                 build_config,
                                 dead_code_graph,
-                                dependency_graph,
                                 mode: Mode::NonAbi,
                                 opts,
                             }),
@@ -935,38 +848,47 @@ impl TypedExpression {
     #[allow(clippy::too_many_arguments)]
     fn type_check_struct_expression(
         span: Span,
-        struct_name: Ident,
+        call_path: CallPath,
         fields: Vec<StructExpressionField>,
         namespace: crate::semantic_analysis::NamespaceRef,
         crate_namespace: NamespaceRef,
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
         let mut errors = vec![];
+        let module_result = namespace
+            .find_module_relative(&call_path.prefixes)
+            .ok(&mut warnings, &mut errors);
         let mut typed_fields_buf = vec![];
-
-        let definition: TypedStructDeclaration =
-            match namespace.clone().get_symbol(&struct_name).value {
+        let definition = match module_result {
+            Some(module) => match module.clone().get_symbol(&call_path.suffix).value {
                 Some(TypedDeclaration::StructDeclaration(st)) => st,
                 Some(_) => {
                     errors.push(CompileError::DeclaredNonStructAsStruct {
-                        name: struct_name.clone(),
+                        name: call_path.suffix.clone(),
                         span,
                     });
                     return err(warnings, errors);
                 }
                 None => {
                     errors.push(CompileError::StructNotFound {
-                        name: struct_name.clone(),
+                        name: call_path.suffix.clone(),
                         span,
                     });
                     return err(warnings, errors);
                 }
-            };
+            },
+            None => {
+                errors.push(CompileError::StructNotFound {
+                    name: call_path.suffix,
+                    span,
+                });
+                return err(warnings, errors);
+            }
+        };
         // if this is a generic struct, i.e. it has some type
         // parameters, monomorphize it before unifying the
         // types
@@ -1011,7 +933,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 }),
@@ -1066,7 +987,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
@@ -1081,7 +1001,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -1141,7 +1060,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
@@ -1171,7 +1089,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 }),
@@ -1236,7 +1153,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
@@ -1251,7 +1167,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -1308,16 +1223,15 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
         let mut errors = vec![];
-        // The first step is to determine if the call path refers to a module or an enum.
+        // The first step is to determine if the call path refers to a module, enum, or function.
         // We could rely on the capitalization convention, where modules are lowercase
         // and enums are uppercase, but this is not robust in the long term.
-        // Instead, we try to resolve both paths.
-        // If only one exists, then we use that one. Otherwise, if both exist, it is
+        // Instead, we try to resolve all paths.
+        // If only one exists, then we use that one. Otherwise, if more than one exist, it is
         // an ambiguous reference error.
         let mut probe_warnings = Vec::new();
         let mut probe_errors = Vec::new();
@@ -1335,60 +1249,91 @@ impl TypedExpression {
         };
 
         // now we can see if this thing is a symbol (typed declaration) or reference to an
-        // enum instantiation
-        let this_thing: Either<TypedDeclaration, TypedExpression> =
-            match (module_result, enum_module_combined_result) {
-                (Some(_module), Some(_enum_res)) => {
-                    errors.push(CompileError::AmbiguousPath { span });
-                    return err(warnings, errors);
-                }
-                (Some(module), None) => match module.get_symbol(&call_path.suffix).value {
-                    Some(decl) => Either::Left(decl),
-                    None => {
-                        errors.push(CompileError::SymbolNotFound {
-                            name: call_path.suffix.as_str().to_string(),
-                            span: call_path.suffix.span().clone(),
+        // enum instantiation, and if it is not either of those things, then it might be a
+        // function application
+        let exp: TypedExpression = match (module_result, enum_module_combined_result) {
+            (Some(_module), Some(_enum_res)) => {
+                errors.push(CompileError::AmbiguousPath { span });
+                return err(warnings, errors);
+            }
+            (Some(module), None) => match module.get_symbol(&call_path.suffix).value {
+                Some(decl) => match decl {
+                    TypedDeclaration::EnumDeclaration(enum_decl) => {
+                        check!(
+                            instantiate_enum(
+                                enum_decl,
+                                call_path.suffix,
+                                args,
+                                namespace,
+                                crate_namespace,
+                                self_type,
+                                build_config,
+                                dead_code_graph,
+                                opts,
+                            ),
+                            return err(warnings, errors),
+                            warnings,
+                            errors
+                        )
+                    }
+                    TypedDeclaration::FunctionDeclaration(func_decl) => check!(
+                        instantiate_function_application(
+                            func_decl,
+                            call_path,
+                            args,
+                            namespace,
+                            crate_namespace,
+                            self_type,
+                            build_config,
+                            dead_code_graph,
+                            opts,
+                        ),
+                        return err(warnings, errors),
+                        warnings,
+                        errors
+                    ),
+                    a => {
+                        errors.push(CompileError::NotAnEnum {
+                            name: call_path.friendly_name(),
+                            span,
+                            actually: a.friendly_name().to_string(),
                         });
                         return err(warnings, errors);
                     }
                 },
-                (None, Some(enum_decl)) => Either::Right(check!(
-                    instantiate_enum(
-                        enum_decl,
-                        call_path.suffix,
-                        args,
-                        namespace,
-                        crate_namespace,
-                        self_type,
-                        build_config,
-                        dead_code_graph,
-                        dependency_graph,
-                        opts,
-                    ),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                )),
-                (None, None) => {
+                None => {
                     errors.push(CompileError::SymbolNotFound {
-                        span,
                         name: call_path.suffix.as_str().to_string(),
+                        span: call_path.suffix.span().clone(),
                     });
                     return err(warnings, errors);
                 }
-            };
-
-        let exp = match this_thing {
-            Either::Left(_) => {
-                errors.push(CompileError::Unimplemented(
-                    "Unable to refer to declarations in other modules directly. Try \
-                     importing it instead.",
+            },
+            (None, Some(enum_decl)) => check!(
+                instantiate_enum(
+                    enum_decl,
+                    call_path.suffix,
+                    args,
+                    namespace,
+                    crate_namespace,
+                    self_type,
+                    build_config,
+                    dead_code_graph,
+                    opts,
+                ),
+                return err(warnings, errors),
+                warnings,
+                errors
+            ),
+            (None, None) => {
+                errors.push(CompileError::SymbolNotFound {
                     span,
-                ));
+                    name: call_path.suffix.as_str().to_string(),
+                });
                 return err(warnings, errors);
             }
-            Either::Right(expr) => expr,
         };
+
         ok(exp, warnings, errors)
     }
 
@@ -1402,7 +1347,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
@@ -1423,7 +1367,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -1473,7 +1416,6 @@ impl TypedExpression {
                     build_config,
                     dead_code_graph,
                     mode: Mode::ImplAbiFn,
-                    dependency_graph,
                     opts,
                 }),
                 return err(warnings, errors),
@@ -1510,7 +1452,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         if contents.is_empty() {
@@ -1544,7 +1485,6 @@ impl TypedExpression {
                         self_type,
                         build_config,
                         dead_code_graph,
-                        dependency_graph,
                         mode: Mode::NonAbi,
                         opts,
                     }),
@@ -1605,7 +1545,6 @@ impl TypedExpression {
             self_type,
             build_config,
             dead_code_graph,
-            dependency_graph,
             opts,
             ..
         } = arguments;
@@ -1622,7 +1561,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -1645,7 +1583,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 }),
@@ -1676,9 +1613,9 @@ impl TypedExpression {
                         Ident::new_with_override("ops", span.clone()),
                     ],
                     suffix: Ident::new_with_override("index", span.clone()),
+                    is_absolute: true,
                 },
                 type_name: None,
-                is_absolute: true,
             };
             type_check_method_application(
                 method_name,
@@ -1689,7 +1626,6 @@ impl TypedExpression {
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 opts,
             )
         }
@@ -1712,7 +1648,6 @@ impl TypedExpression {
         self_type: TypeId,
         build_config: &BuildConfig,
         dead_code_graph: &mut ControlFlowGraph,
-        dependency_graph: &mut HashMap<String, HashSet<String>>,
         opts: TCOpts,
     ) -> CompileResult<TypedExpression> {
         let mut warnings = vec![];
@@ -1731,7 +1666,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 };
@@ -1795,7 +1729,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 };
@@ -1888,7 +1821,6 @@ impl TypedExpression {
                     self_type,
                     build_config,
                     dead_code_graph,
-                    dependency_graph,
                     mode: Mode::NonAbi,
                     opts,
                 };
@@ -1947,6 +1879,62 @@ impl TypedExpression {
                 ok(exp, warnings, errors)
             }
         }
+    }
+
+    fn type_check_size_of_val(
+        arguments: TypeCheckArguments<'_, Expression>,
+        span: Span,
+    ) -> CompileResult<TypedExpression> {
+        let mut warnings = vec![];
+        let mut errors = vec![];
+        let exp = check!(
+            TypedExpression::type_check(arguments),
+            return err(warnings, errors),
+            warnings,
+            errors
+        );
+        let exp = TypedExpression {
+            expression: TypedExpressionVariant::SizeOf {
+                variant: SizeOfVariant::Val(Box::new(exp)),
+            },
+            return_type: crate::type_engine::insert_type(TypeInfo::UnsignedInteger(
+                IntegerBits::SixtyFour,
+            )),
+            is_constant: IsConstant::No,
+            span,
+        };
+        ok(exp, warnings, errors)
+    }
+
+    fn type_check_size_of_type(
+        arguments: TypeCheckArguments<'_, (TypeInfo, Span)>,
+        span: Span,
+    ) -> CompileResult<TypedExpression> {
+        let warnings = vec![];
+        let mut errors = vec![];
+        let TypeCheckArguments {
+            checkee: (type_name, type_span),
+            self_type,
+            namespace,
+            ..
+        } = arguments;
+        let type_id = namespace
+            .resolve_type_with_self(type_name, self_type)
+            .unwrap_or_else(|_| {
+                errors.push(CompileError::UnknownType { span: type_span });
+                insert_type(TypeInfo::ErrorRecovery)
+            });
+        let exp = TypedExpression {
+            expression: TypedExpressionVariant::SizeOf {
+                variant: SizeOfVariant::Type(type_id),
+            },
+            return_type: crate::type_engine::insert_type(TypeInfo::UnsignedInteger(
+                IntegerBits::SixtyFour,
+            )),
+            is_constant: IsConstant::No,
+            span,
+        };
+        ok(exp, warnings, errors)
     }
 
     fn resolve_numeric_literal(
@@ -2070,7 +2058,6 @@ mod tests {
             generated_names: Arc::new(Mutex::new(vec![])),
         };
         let mut dead_code_graph: ControlFlowGraph = Default::default();
-        let mut dependency_graph = HashMap::new();
 
         TypedExpression::type_check(TypeCheckArguments {
             checkee: expr,
@@ -2081,7 +2068,6 @@ mod tests {
             self_type,
             build_config: &build_config,
             dead_code_graph: &mut dead_code_graph,
-            dependency_graph: &mut dependency_graph,
             mode: Mode::NonAbi,
             opts: Default::default(),
         })
