@@ -5,7 +5,7 @@ use crate::parse_tree::MethodName;
 use crate::parser::{Rule, SwayParser};
 use crate::semantic_analysis::TCOpts;
 use pest::Parser;
-use std::collections::{HashMap, VecDeque};
+use std::collections::VecDeque;
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn type_check_method_application(
@@ -17,7 +17,6 @@ pub(crate) fn type_check_method_application(
     self_type: TypeId,
     build_config: &BuildConfig,
     dead_code_graph: &mut ControlFlowGraph,
-    dependency_graph: &mut HashMap<String, HashSet<String>>,
     opts: TCOpts,
 ) -> CompileResult<TypedExpression> {
     let mut warnings = vec![];
@@ -34,7 +33,6 @@ pub(crate) fn type_check_method_application(
                 self_type,
                 build_config,
                 dead_code_graph,
-                dependency_graph,
                 mode: Mode::NonAbi,
                 opts,
             }),
@@ -48,7 +46,6 @@ pub(crate) fn type_check_method_application(
         MethodName::FromType {
             ref type_name,
             ref call_path,
-            is_absolute,
         } => {
             let ty = match type_name {
                 Some(name) => {
@@ -63,7 +60,7 @@ pub(crate) fn type_check_method_application(
                     .map(|x| x.return_type)
                     .unwrap_or_else(|| insert_type(TypeInfo::Unknown)),
             };
-            let from_module = if is_absolute {
+            let from_module = if call_path.is_absolute {
                 Some(crate_namespace)
             } else {
                 None
@@ -156,6 +153,7 @@ pub(crate) fn type_check_method_application(
                     name: CallPath {
                         prefixes: vec![],
                         suffix: method_name,
+                        is_absolute: false,
                     },
                     arguments: args_and_names,
                     function_body: method.body.clone(),
@@ -182,7 +180,6 @@ pub(crate) fn type_check_method_application(
                                 crate_namespace,
                                 self_type,
                                 dead_code_graph,
-                                dependency_graph,
                                 opts,
                             ),
                             return err(warnings, errors),
@@ -257,7 +254,6 @@ pub(crate) fn type_check_method_application(
                                 crate_namespace,
                                 self_type,
                                 dead_code_graph,
-                                dependency_graph,
                                 opts,
                             ),
                             return err(warnings, errors),
@@ -293,7 +289,6 @@ fn re_parse_expression(
     crate_namespace: NamespaceRef,
     self_type: TypeId,
     dead_code_graph: &mut ControlFlowGraph,
-    dependency_graph: &mut HashMap<String, HashSet<String>>,
     opts: TCOpts,
 ) -> CompileResult<TypedExpression> {
     let mut warnings = vec![];
@@ -345,7 +340,6 @@ fn re_parse_expression(
             self_type,
             build_config,
             dead_code_graph,
-            dependency_graph,
             mode: Mode::NonAbi,
             opts,
         }),
