@@ -14,15 +14,12 @@ use crate::{
 
 use sway_types::span::Span;
 
-use std::collections::{HashMap, HashSet};
-
 pub(crate) fn implementation_of_trait(
     impl_trait: ImplTrait,
     namespace: crate::semantic_analysis::NamespaceRef,
     crate_namespace: NamespaceRef,
     build_config: &BuildConfig,
     dead_code_graph: &mut ControlFlowGraph,
-    dependency_graph: &mut HashMap<String, HashSet<String>>,
     opts: TCOpts,
 ) -> CompileResult<TypedDeclaration> {
     let mut errors = vec![];
@@ -74,7 +71,6 @@ pub(crate) fn implementation_of_trait(
                     type_implementing_for_id,
                     &type_implementing_for_span,
                     Mode::NonAbi,
-                    dependency_graph,
                     opts,
                 ),
                 return err(warnings, errors),
@@ -135,7 +131,6 @@ pub(crate) fn implementation_of_trait(
                     type_implementing_for_id,
                     &type_implementing_for_span,
                     Mode::ImplAbiFn,
-                    dependency_graph,
                     opts,
                 ),
                 return err(warnings, errors),
@@ -199,7 +194,6 @@ fn type_check_trait_implementation(
     type_implementing_for: TypeId,
     type_implementing_for_span: &Span,
     mode: Mode,
-    dependency_graph: &mut HashMap<String, HashSet<String>>,
     opts: TCOpts,
 ) -> CompileResult<Vec<TypedFunctionDeclaration>> {
     let mut functions_buf: Vec<TypedFunctionDeclaration> = vec![];
@@ -229,7 +223,6 @@ fn type_check_trait_implementation(
                 build_config,
                 dead_code_graph,
                 mode,
-                dependency_graph,
                 opts,
             }),
             continue,
@@ -358,6 +351,7 @@ fn type_check_trait_implementation(
         CallPath {
             prefixes: vec![],
             suffix: trait_name.clone(),
+            is_absolute: false,
         },
         match resolve_type(type_implementing_for, type_implementing_for_span) {
             Ok(o) => o,
@@ -385,7 +379,6 @@ fn type_check_trait_implementation(
                 build_config,
                 dead_code_graph,
                 mode,
-                dependency_graph,
                 opts,
             }),
             continue,
