@@ -372,8 +372,9 @@ impl Dependencies {
                 fields,
                 ..
             } => {
-                self.deps
-                    .insert(DependentSymbol::Symbol(struct_name.as_str().to_string()));
+                self.deps.insert(DependentSymbol::Symbol(
+                    struct_name.suffix.as_str().to_string(),
+                ));
                 self.gather_from_iter(fields.iter(), |deps, field| {
                     deps.gather_from_expr(&field.value)
                 })
@@ -419,6 +420,8 @@ impl Dependencies {
             Expression::TupleIndex { prefix, .. } => self.gather_from_expr(prefix),
             Expression::DelayedMatchTypeResolution { .. } => self,
             Expression::StorageAccess { .. } => self,
+            Expression::SizeOfVal { exp, .. } => self.gather_from_expr(exp),
+            Expression::SizeOfType { .. } => self,
         }
     }
 
