@@ -14,8 +14,9 @@
 //! }
 //! ```
 
-use crate::{context::Context, irtype::Type, value::Value};
 use sway_types::ident::Ident;
+
+use crate::{context::Context, irtype::Type, metadata::MetadataIndex, value::Value};
 
 /// A wrapper around an [ECS](https://github.com/fitzgen/generational-arena) handle into the
 /// [`Context`].
@@ -41,6 +42,7 @@ pub struct AsmInstruction {
     pub name: Ident,
     pub args: Vec<Ident>,
     pub immediate: Option<Ident>,
+    pub span_md_idx: Option<MetadataIndex>,
 }
 
 impl AsmBlock {
