@@ -8,18 +8,17 @@
 //! this should be addressed in the future, perhaps by using compiler intrinsic calls instead of
 //! the ASM blocks where possible.
 
+use sway_types::ident::Ident;
+
 use crate::{
     asm::{AsmArg, AsmBlock, AsmInstruction},
     block::Block,
     context::Context,
     function::Function,
     irtype::{Aggregate, Type},
+    metadata::MetadataIndex,
     pointer::Pointer,
     value::Value,
-};
-use sway_types::{
-    ident::Ident,
-    state::{StateIndex, StateSlot},
 };
 
 #[derive(Debug, Clone)]
@@ -66,21 +65,14 @@ pub enum Instruction {
     },
     /// Read a value from a memory pointer.
     Load(Pointer),
+    /// No-op, handy as a placeholder instruction.
+    Nop,
     /// Choose a value from a list depending on the preceding block.
     Phi(Vec<(Block, Value)>),
     /// Return from a function.
     Ret(Value, Type),
     /// Write a value to a memory pointer.
-    Store {
-        ptr: Pointer,
-        stored_val: Value,
-    },
-    StateStore {
-        slot: StateIndex,
-    },
-    StateLoad {
-        slot: StateIndex,
-    },
+    Store { ptr: Pointer, stored_val: Value },
 }
 
 impl Instruction {
@@ -111,27 +103,28 @@ impl Instruction {
             Instruction::InsertElement { .. } => None,
             Instruction::InsertValue { .. } => None,
             Instruction::Store { .. } => None,
-            Instruction::StateStore { .. } | Instruction::StateLoad { .. } => {
-                todo!("should this be None?")
-            }
+
+            // No-op is also no-type.
+            Instruction::Nop => None,
         }
     }
 
     /// Some [`Instruction`]s may have struct arguments.  Return it if so for this instruction.
     pub fn get_aggregate(&self, context: &Context) -> Option<Aggregate> {
-        use Instruction::*;
         match self {
-            GetPointer(ptr) | Load(ptr) => match ptr.get_type(context) {
+            Instruction::GetPointer(ptr) | Instruction::Load(ptr) => match ptr.get_type(context) {
                 Type::Array(aggregate) => Some(*aggregate),
                 Type::Struct(aggregate) => Some(*aggregate),
                 _otherwise => None,
             },
-            ExtractElement { ty, .. } => ty.get_elem_type(context).and_then(|ty| match ty {
-                Type::Array(nested_aggregate) => Some(nested_aggregate),
-                Type::Struct(nested_aggregate) => Some(nested_aggregate),
-                _otherwise => None,
-            }),
-            ExtractValue { ty, indices, .. } => {
+            Instruction::ExtractElement { ty, .. } => {
+                ty.get_elem_type(context).and_then(|ty| match ty {
+                    Type::Array(nested_aggregate) => Some(nested_aggregate),
+                    Type::Struct(nested_aggregate) => Some(nested_aggregate),
+                    _otherwise => None,
+                })
+            }
+            Instruction::ExtractValue { ty, indices, .. } => {
                 // This array is a field in a struct or element in an array.
                 ty.get_field_type(context, indices).and_then(|ty| match ty {
                     Type::Array(nested_aggregate) => Some(nested_aggregate),
@@ -139,18 +132,9 @@ impl Instruction {
                     _otherwise => None,
                 })
             }
-            StateStore { slot } => todo!("get this state type out of the context?"),
-            StateLoad { slot } => todo!("get this state type out of the context?"),
+
             // Unknown aggregate instruction.  Adding these as we come across them...
-            AsmBlock(_, _)
-            | Branch(_)
-            | Call(_, _)
-            | ConditionalBranch { .. }
-            | InsertElement { .. }
-            | InsertValue { .. }
-            | Phi(_)
-            | Ret(_, _)
-            | Store { .. } => None,
+            _otherwise => None,
         }
     }
 
@@ -196,13 +180,12 @@ impl Instruction {
             }
             Instruction::ExtractValue { aggregate, .. } => replace(aggregate),
             Instruction::Load(_) => (),
+            Instruction::Nop => (),
             Instruction::Phi(pairs) => pairs.iter_mut().for_each(|(_, val)| replace(val)),
             Instruction::Ret(ret_val, _) => replace(ret_val),
             Instruction::Store { stored_val, .. } => {
                 replace(stored_val);
             }
-            Instruction::StateStore { .. } => todo!("compile to state store"),
-            Instruction::StateLoad { .. } => todo!("compile to state load"),
         }
     }
 }
@@ -264,6 +247,7 @@ impl<'a> InstructionInserter<'a> {
         args: Vec<AsmArg>,
         body: Vec<AsmInstruction>,
         return_name: Option<Ident>,
+        span_md_idx: Option<MetadataIndex>,
     ) -> Value {
         let asm = AsmBlock::new(
             self.context,
@@ -271,17 +255,29 @@ impl<'a> InstructionInserter<'a> {
             body,
             return_name,
         );
-        self.asm_block_from_asm(asm, args)
+        self.asm_block_from_asm(asm, args, span_md_idx)
     }
 
-    pub fn asm_block_from_asm(self, asm: AsmBlock, args: Vec<AsmArg>) -> Value {
-        let asm_val = Value::new_instruction(self.context, Instruction::AsmBlock(asm, args));
+    pub fn asm_block_from_asm(
+        self,
+        asm: AsmBlock,
+        args: Vec<AsmArg>,
+        span_md_idx: Option<MetadataIndex>,
+    ) -> Value {
+        let asm_val =
+            Value::new_instruction(self.context, Instruction::AsmBlock(asm, args), span_md_idx);
         self.context.blocks[self.block.0].instructions.push(asm_val);
         asm_val
     }
 
-    pub fn branch(self, to_block: Block, phi_value: Option<Value>) -> Value {
-        let br_val = Value::new_instruction(self.context, Instruction::Branch(to_block));
+    pub fn branch(
+        self,
+        to_block: Block,
+        phi_value: Option<Value>,
+        span_md_idx: Option<MetadataIndex>,
+    ) -> Value {
+        let br_val =
+            Value::new_instruction(self.context, Instruction::Branch(to_block), span_md_idx);
         phi_value
             .into_iter()
             .for_each(|pv| to_block.add_phi(self.context, self.block, pv));
@@ -289,9 +285,17 @@ impl<'a> InstructionInserter<'a> {
         br_val
     }
 
-    pub fn call(self, function: Function, args: &[Value]) -> Value {
-        let call_val =
-            Value::new_instruction(self.context, Instruction::Call(function, args.to_vec()));
+    pub fn call(
+        self,
+        function: Function,
+        args: &[Value],
+        span_md_idx: Option<MetadataIndex>,
+    ) -> Value {
+        let call_val = Value::new_instruction(
+            self.context,
+            Instruction::Call(function, args.to_vec()),
+            span_md_idx,
+        );
         self.context.blocks[self.block.0]
             .instructions
             .push(call_val);
@@ -304,6 +308,7 @@ impl<'a> InstructionInserter<'a> {
         true_block: Block,
         false_block: Block,
         phi_value: Option<Value>,
+        span_md_idx: Option<MetadataIndex>,
     ) -> Value {
         let cbr_val = Value::new_instruction(
             self.context,
@@ -312,6 +317,7 @@ impl<'a> InstructionInserter<'a> {
                 true_block,
                 false_block,
             },
+            span_md_idx,
         );
         phi_value.into_iter().for_each(|pv| {
             true_block.add_phi(self.context, self.block, pv);
@@ -321,7 +327,13 @@ impl<'a> InstructionInserter<'a> {
         cbr_val
     }
 
-    pub fn extract_element(self, array: Value, ty: Aggregate, index_val: Value) -> Value {
+    pub fn extract_element(
+        self,
+        array: Value,
+        ty: Aggregate,
+        index_val: Value,
+        span_md_idx: Option<MetadataIndex>,
+    ) -> Value {
         let extract_element_val = Value::new_instruction(
             self.context,
             Instruction::ExtractElement {
@@ -329,6 +341,7 @@ impl<'a> InstructionInserter<'a> {
                 ty,
                 index_val,
             },
+            span_md_idx,
         );
         self.context.blocks[self.block.0]
             .instructions
@@ -336,7 +349,13 @@ impl<'a> InstructionInserter<'a> {
         extract_element_val
     }
 
-    pub fn extract_value(self, aggregate: Value, ty: Aggregate, indices: Vec<u64>) -> Value {
+    pub fn extract_value(
+        self,
+        aggregate: Value,
+        ty: Aggregate,
+        indices: Vec<u64>,
+        span_md_idx: Option<MetadataIndex>,
+    ) -> Value {
         let extract_value_val = Value::new_instruction(
             self.context,
             Instruction::ExtractValue {
@@ -344,6 +363,7 @@ impl<'a> InstructionInserter<'a> {
                 ty,
                 indices,
             },
+            span_md_idx,
         );
         self.context.blocks[self.block.0]
             .instructions
@@ -351,8 +371,9 @@ impl<'a> InstructionInserter<'a> {
         extract_value_val
     }
 
-    pub fn get_ptr(self, ptr: Pointer) -> Value {
-        let get_ptr_val = Value::new_instruction(self.context, Instruction::GetPointer(ptr));
+    pub fn get_ptr(self, ptr: Pointer, span_md_idx: Option<MetadataIndex>) -> Value {
+        let get_ptr_val =
+            Value::new_instruction(self.context, Instruction::GetPointer(ptr), span_md_idx);
         self.context.blocks[self.block.0]
             .instructions
             .push(get_ptr_val);
@@ -365,6 +386,7 @@ impl<'a> InstructionInserter<'a> {
         ty: Aggregate,
         value: Value,
         index_val: Value,
+        span_md_idx: Option<MetadataIndex>,
     ) -> Value {
         let insert_val = Value::new_instruction(
             self.context,
@@ -374,6 +396,7 @@ impl<'a> InstructionInserter<'a> {
                 value,
                 index_val,
             },
+            span_md_idx,
         );
         self.context.blocks[self.block.0]
             .instructions
@@ -387,6 +410,7 @@ impl<'a> InstructionInserter<'a> {
         ty: Aggregate,
         value: Value,
         indices: Vec<u64>,
+        span_md_idx: Option<MetadataIndex>,
     ) -> Value {
         let insert_val = Value::new_instruction(
             self.context,
@@ -396,6 +420,7 @@ impl<'a> InstructionInserter<'a> {
                 value,
                 indices,
             },
+            span_md_idx,
         );
         self.context.blocks[self.block.0]
             .instructions
@@ -403,23 +428,38 @@ impl<'a> InstructionInserter<'a> {
         insert_val
     }
 
-    pub fn load(self, ptr: Pointer) -> Value {
-        let load_val = Value::new_instruction(self.context, Instruction::Load(ptr));
+    pub fn load(self, ptr: Pointer, span_md_idx: Option<MetadataIndex>) -> Value {
+        let load_val = Value::new_instruction(self.context, Instruction::Load(ptr), span_md_idx);
         self.context.blocks[self.block.0]
             .instructions
             .push(load_val);
         load_val
     }
 
-    pub fn ret(self, value: Value, ty: Type) -> Value {
-        let ret_val = Value::new_instruction(self.context, Instruction::Ret(value, ty));
+    pub fn nop(self) -> Value {
+        let nop_val = Value::new_instruction(self.context, Instruction::Nop, None);
+        self.context.blocks[self.block.0].instructions.push(nop_val);
+        nop_val
+    }
+
+    pub fn ret(self, value: Value, ty: Type, span_md_idx: Option<MetadataIndex>) -> Value {
+        let ret_val =
+            Value::new_instruction(self.context, Instruction::Ret(value, ty), span_md_idx);
         self.context.blocks[self.block.0].instructions.push(ret_val);
         ret_val
     }
 
-    pub fn store(self, ptr: Pointer, stored_val: Value) -> Value {
-        let store_val =
-            Value::new_instruction(self.context, Instruction::Store { ptr, stored_val });
+    pub fn store(
+        self,
+        ptr: Pointer,
+        stored_val: Value,
+        span_md_idx: Option<MetadataIndex>,
+    ) -> Value {
+        let store_val = Value::new_instruction(
+            self.context,
+            Instruction::Store { ptr, stored_val },
+            span_md_idx,
+        );
         self.context.blocks[self.block.0]
             .instructions
             .push(store_val);
