@@ -1,17 +1,17 @@
 //! A parser for the printed IR, useful mostly for testing.
 
-use crate::context::Context;
+use crate::{context::Context, error::IrError};
 
 // -------------------------------------------------------------------------------------------------
 /// Parse a string produced by [`crate::printer::to_string`] into a new [`Context`].
-pub fn parse(input: &str) -> Result<Context, String> {
+pub fn parse(input: &str) -> Result<Context, IrError> {
     let irmod = ir_builder::parser::ir_descrs(input).map_err(|err| {
         let found = if input.len() - err.location.offset <= 20 {
             &input[err.location.offset..]
         } else {
             &input[err.location.offset..][..20]
         };
-        format!("parse failed: {}, found: {}", err, found)
+        IrError::ParseFailure(err.to_string(), found.into())
     })?;
     ir_builder::build_context(irmod)
 }
@@ -19,7 +19,11 @@ pub fn parse(input: &str) -> Result<Context, String> {
 // -------------------------------------------------------------------------------------------------
 
 mod ir_builder {
+    use std::path::PathBuf;
+
     use sway_types::{ident::Ident, span::Span};
+
+    type MdIdxRef = u64;
 
     peg::parser! {
         pub(in crate::parser) grammar parser() for str {
@@ -29,11 +33,11 @@ mod ir_builder {
                 }
 
             rule script() -> IrAstModule
-                = "script" _ name:id() "{" _ fn_decls:fn_decl()* "}" _ {
+                = "script" _ "{" _ fn_decls:fn_decl()* "}" _ metadata:metadata_decl()* {
                     IrAstModule {
-                        name,
                         kind: crate::module::Kind::Script,
-                        fn_decls
+                        fn_decls,
+                        metadata
                     }
                 }
 
@@ -51,9 +55,9 @@ mod ir_builder {
                     }
                 }
 
-            rule fn_arg() -> (IrAstTy, String)
-                = name:id() ":" _ ty:ast_ty() {
-                    (ty, name)
+            rule fn_arg() -> (IrAstTy, String, Option<MdIdxRef>)
+                = name:id() mdi:metadata_idx()? ":" _ ty:ast_ty() {
+                    (ty, name, mdi)
                 }
 
             rule fn_local() -> (IrAstTy, String, bool, Option<IrAstOperation>)
@@ -75,16 +79,27 @@ mod ir_builder {
                 }
 
             rule instr_decl() -> IrAstInstruction
-                = value_name:value_assign()? op:operation() {
+                = value_name:value_assign()? op:operation() meta_idx:comma_metadata_idx()? {
                     IrAstInstruction {
                         value_name,
                         op,
+                        meta_idx,
                     }
                 }
 
             rule value_assign() -> String
                 = name:id() "=" _ {
                     name
+                }
+
+            rule metadata_idx() -> MdIdxRef
+                = "!" i:decimal() {
+                    i
+                }
+
+            rule comma_metadata_idx() -> MdIdxRef
+                = "," _ mdi:metadata_idx() {
+                    mdi
                 }
 
             rule operation() -> IrAstOperation
@@ -99,15 +114,16 @@ mod ir_builder {
                 / op_insert_element()
                 / op_insert_value()
                 / op_load()
+                / op_nop()
                 / op_phi()
                 / op_ret()
                 / op_store()
 
             rule op_asm() -> IrAstOperation
-                = "asm" _ "(" _ args:(asm_arg() ** comma()) ")" _ ret:asm_ret()? "{" _
+                = "asm" _ "(" _ args:(asm_arg() ** comma()) ")" _ ret:asm_ret()? meta_idx:comma_metadata_idx()? "{" _
                     ops:asm_op()*
                 "}" _ {
-                    IrAstOperation::Asm(args, ret, ops)
+                    IrAstOperation::Asm(args, ret, ops, meta_idx)
                 }
 
             rule op_branch() -> IrAstOperation
@@ -160,6 +176,11 @@ mod ir_builder {
                     IrAstOperation::Load(src)
                 }
 
+            rule op_nop() -> IrAstOperation
+                = "nop" _ {
+                    IrAstOperation::Nop
+                }
+
             rule op_phi() -> IrAstOperation
                 = "phi" _ "(" _ pairs:((bl:id() ":" _ vn:id() { (bl, vn) }) ** comma()) ")" _ {
                     IrAstOperation::Phi(pairs)
@@ -194,11 +215,12 @@ mod ir_builder {
                 }
 
             rule asm_op() -> IrAstAsmOp
-                = name:id_id() args:asm_op_arg()* imm:asm_op_arg_imm()? {
+                = name:id_id() args:asm_op_arg()* imm:asm_op_arg_imm()? meta_idx:comma_metadata_idx()? {
                     IrAstAsmOp {
                         name,
                         args,
-                        imm
+                        imm,
+                        meta_idx
                     }
                 }
 
@@ -215,7 +237,15 @@ mod ir_builder {
                     })
                 }
 
-            rule constant() -> IrAstConstValue
+            rule constant() -> IrAstConst
+                = value:constant_value() meta_idx:metadata_idx()? {
+                    IrAstConst {
+                        value,
+                        meta_idx
+                    }
+                }
+
+            rule constant_value() -> IrAstConstValue
                 = "()" _ { IrAstConstValue::Unit }
                 / "true" _ { IrAstConstValue::Bool(true) }
                 / "false" _ { IrAstConstValue::Bool(false) }
@@ -247,7 +277,7 @@ mod ir_builder {
             rule array_const() -> IrAstConstValue
                 = "[" _ els:(field_or_element_const() ++ comma()) "]" _ {
                     let el_ty = els[0].0.clone();
-                    let els = els.into_iter().map(|(_, cv)| cv).collect();
+                    let els = els.into_iter().map(|(_, cv)| cv).collect::<Vec<_>>();
                     IrAstConstValue::Array(el_ty, els)
                 }
 
@@ -256,12 +286,12 @@ mod ir_builder {
                     IrAstConstValue::Struct(flds)
                 }
 
-            rule field_or_element_const() -> (IrAstTy, IrAstConstValue)
+            rule field_or_element_const() -> (IrAstTy, IrAstConst)
                 = ty:ast_ty() cv:constant() {
                     (ty, cv)
                 }
                 / ty:ast_ty() "undef" _ {
-                    (ty.clone(), IrAstConstValue::Undef(ty))
+                    (ty.clone(), IrAstConst { value: IrAstConstValue::Undef(ty), meta_idx: None })
                 }
 
             rule ast_ty() -> IrAstTy
@@ -300,6 +330,19 @@ mod ir_builder {
                         span: pest::Span::new(id.into(), 0, id.len()).unwrap(),
                         path: None,
                     })
+                }
+
+            rule metadata_decl() -> (MdIdxRef, IrMetadatum)
+                = "!" idx:decimal() "=" _ item:metadata_item() {
+                    (idx, item)
+                }
+
+            rule metadata_item() -> IrMetadatum
+                = "filepath" _ ['"'] path:$(([^ '"' | '\\'] / ['\\'] ['\\' | '"' ])+) ['"'] _ {
+                    IrMetadatum::FilePath(PathBuf::from(path))
+                }
+                / "span" _ "!" li:decimal() s:decimal() e:decimal() {
+                    IrMetadatum::Span { loc_idx: li, start: s as usize, end: e as usize }
                 }
 
             rule id_char0()
@@ -341,8 +384,11 @@ mod ir_builder {
         block::Block,
         constant::Constant,
         context::Context,
+        error::IrError,
         function::Function,
+        instruction::Instruction,
         irtype::{Aggregate, Type},
+        metadata::{MetadataIndex, Metadatum},
         module::{Kind, Module},
         pointer::Pointer,
         value::Value,
@@ -350,15 +396,15 @@ mod ir_builder {
 
     #[derive(Debug)]
     pub(super) struct IrAstModule {
-        name: String,
         kind: Kind,
         fn_decls: Vec<IrAstFnDecl>,
+        metadata: Vec<(MdIdxRef, IrMetadatum)>,
     }
 
     #[derive(Debug)]
     struct IrAstFnDecl {
         name: String,
-        args: Vec<(IrAstTy, String)>,
+        args: Vec<(IrAstTy, String, Option<MdIdxRef>)>,
         ret_type: IrAstTy,
         locals: Vec<(IrAstTy, String, bool, Option<IrAstOperation>)>,
         blocks: Vec<IrAstBlock>,
@@ -374,6 +420,7 @@ mod ir_builder {
     struct IrAstInstruction {
         value_name: Option<String>,
         op: IrAstOperation,
+        meta_idx: Option<MdIdxRef>,
     }
 
     #[derive(Debug)]
@@ -382,20 +429,28 @@ mod ir_builder {
             Vec<(Ident, Option<IrAstAsmArgInit>)>,
             Option<Ident>,
             Vec<IrAstAsmOp>,
+            Option<MdIdxRef>,
         ),
         Br(String),
         Call(String, Vec<String>),
         Cbr(String, String, String),
-        Const(IrAstConstValue),
+        Const(IrAstConst),
         ExtractElement(String, IrAstTy, String),
         ExtractValue(String, IrAstTy, Vec<u64>),
         GetPtr(String),
         InsertElement(String, IrAstTy, String, String),
         InsertValue(String, IrAstTy, String, Vec<u64>),
         Load(String),
+        Nop,
         Phi(Vec<(String, String)>),
         Ret(IrAstTy, String),
         Store(String, String),
+    }
+
+    #[derive(Debug)]
+    struct IrAstConst {
+        value: IrAstConstValue,
+        meta_idx: Option<MdIdxRef>,
     }
 
     #[derive(Debug)]
@@ -406,14 +461,14 @@ mod ir_builder {
         B256([u8; 32]),
         Number(u64),
         String(String),
-        Array(IrAstTy, Vec<IrAstConstValue>),
-        Struct(Vec<(IrAstTy, IrAstConstValue)>),
+        Array(IrAstTy, Vec<IrAstConst>),
+        Struct(Vec<(IrAstTy, IrAstConst)>),
     }
 
     #[derive(Debug)]
     enum IrAstAsmArgInit {
         Var(String),
-        Imm(IrAstConstValue),
+        Imm(IrAstConst),
     }
 
     #[derive(Debug)]
@@ -421,6 +476,7 @@ mod ir_builder {
         name: Ident,
         args: Vec<Ident>,
         imm: Option<Ident>,
+        meta_idx: Option<MdIdxRef>,
     }
 
     impl IrAstConstValue {
@@ -436,7 +492,7 @@ mod ir_builder {
                 IrAstConstValue::Number(n) => Constant::new_uint(64, *n),
                 IrAstConstValue::String(s) => Constant::new_string(s.clone()),
                 IrAstConstValue::Array(el_ty, els) => {
-                    let els: Vec<_> = els.iter().map(|cv| cv.as_constant(context)).collect();
+                    let els: Vec<_> = els.iter().map(|cv| cv.value.as_constant(context)).collect();
                     let el_ty = el_ty.to_ir_type(context);
                     let array = Aggregate::new_array(context, el_ty, els.len() as u64);
                     Constant::new_array(&array, els)
@@ -445,29 +501,29 @@ mod ir_builder {
                     // To Make a Constant I need to create an aggregate, which requires a context.
                     let (types, fields): (Vec<_>, Vec<_>) = flds
                         .iter()
-                        .map(|(ty, cv)| (ty.to_ir_type(context), cv.as_constant(context)))
+                        .map(|(ty, cv)| (ty.to_ir_type(context), cv.value.as_constant(context)))
                         .unzip();
-                    let aggregate = Aggregate::new_struct(context, None, types);
+                    let aggregate = Aggregate::new_struct(context, types);
                     Constant::new_struct(&aggregate, fields)
                 }
             }
         }
 
-        fn as_value(&self, context: &mut Context) -> Value {
+        fn as_value(&self, context: &mut Context, span_md_idx: Option<MetadataIndex>) -> Value {
             match self {
                 IrAstConstValue::Undef(_) => unreachable!("Can't convert 'undef' to a value."),
-                IrAstConstValue::Unit => Constant::get_unit(context),
-                IrAstConstValue::Bool(b) => Constant::get_bool(context, *b),
-                IrAstConstValue::B256(bs) => Constant::get_b256(context, *bs),
-                IrAstConstValue::Number(n) => Constant::get_uint(context, 64, *n),
-                IrAstConstValue::String(s) => Constant::get_string(context, s.clone()),
+                IrAstConstValue::Unit => Constant::get_unit(context, span_md_idx),
+                IrAstConstValue::Bool(b) => Constant::get_bool(context, *b, span_md_idx),
+                IrAstConstValue::B256(bs) => Constant::get_b256(context, *bs, span_md_idx),
+                IrAstConstValue::Number(n) => Constant::get_uint(context, 64, *n, span_md_idx),
+                IrAstConstValue::String(s) => Constant::get_string(context, s.clone(), span_md_idx),
                 IrAstConstValue::Array(..) => {
                     let array_const = self.as_constant(context);
-                    Constant::get_array(context, array_const)
+                    Constant::get_array(context, array_const, span_md_idx)
                 }
                 IrAstConstValue::Struct(_) => {
                     let struct_const = self.as_constant(context);
-                    Constant::get_struct(context, struct_const)
+                    Constant::get_struct(context, struct_const, span_md_idx)
                 }
             }
         }
@@ -507,7 +563,7 @@ mod ir_builder {
                 }
                 IrAstTy::Struct(tys) | IrAstTy::Union(tys) => {
                     let tys = tys.iter().map(|ty| ty.to_ir_type(context)).collect();
-                    Aggregate::new_struct(context, None, tys)
+                    Aggregate::new_struct(context, tys)
                 }
                 _otherwise => {
                     unreachable!("Converting non aggregate IR AST type to IR aggregate type.")
@@ -516,29 +572,50 @@ mod ir_builder {
         }
     }
 
+    #[derive(Debug)]
+    enum IrMetadatum {
+        FilePath(PathBuf),
+        Span {
+            loc_idx: MdIdxRef,
+            start: usize,
+            end: usize,
+        },
+    }
+
     // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-    use std::collections::HashMap;
-    use std::iter::FromIterator;
+    use std::{collections::HashMap, iter::FromIterator, sync::Arc};
 
-    pub(super) fn build_context(ir_ast_mod: IrAstModule) -> Result<Context, String> {
+    pub(super) fn build_context(ir_ast_mod: IrAstModule) -> Result<Context, IrError> {
         let mut ctx = Context::default();
-        let module = Module::new(&mut ctx, ir_ast_mod.kind, &ir_ast_mod.name);
+        let module = Module::new(&mut ctx, ir_ast_mod.kind);
+        let md_map = build_metadata_map(&mut ctx, &ir_ast_mod.metadata);
+        let mut unresolved_calls = Vec::new();
         for fn_decl in ir_ast_mod.fn_decls {
-            build_add_fn_decl(&mut ctx, module, fn_decl)?;
+            build_add_fn_decl(&mut ctx, module, fn_decl, &md_map, &mut unresolved_calls)?;
         }
+        resolve_calls(&mut ctx, unresolved_calls)?;
         Ok(ctx)
     }
 
+    #[allow(clippy::type_complexity)]
     fn build_add_fn_decl(
         context: &mut Context,
         module: Module,
         fn_decl: IrAstFnDecl,
-    ) -> Result<(), String> {
-        let args: Vec<(String, Type)> = fn_decl
+        md_map: &HashMap<MdIdxRef, MetadataIndex>,
+        unresolved_calls: &mut Vec<(Block, Value, String, Vec<Value>, Option<MetadataIndex>)>,
+    ) -> Result<(), IrError> {
+        let args: Vec<(String, Type, Option<MetadataIndex>)> = fn_decl
             .args
             .iter()
-            .map(|(ty, name)| (name.into(), ty.to_ir_type(context)))
+            .map(|(ty, name, md_idx)| {
+                (
+                    name.into(),
+                    ty.to_ir_type(context),
+                    md_idx.map(|mdi| md_map.get(&mdi).copied().unwrap()),
+                )
+            })
             .collect();
         let ret_type = fn_decl.ret_type.to_ir_type(context);
         let func = Function::new(
@@ -553,7 +630,7 @@ mod ir_builder {
 
         // Gather all the (new) arg values by name into a map.
         let mut arg_map: HashMap<String, Value> =
-            HashMap::from_iter(args.into_iter().map(|(name, _)| {
+            HashMap::from_iter(args.into_iter().map(|(name, _, _)| {
                 let arg_val = func.get_arg(context, &name).unwrap();
                 (name, arg_val)
             }));
@@ -561,7 +638,7 @@ mod ir_builder {
         for (ty, name, is_mutable, initializer) in fn_decl.locals {
             let initializer = initializer.map(|const_init| {
                 if let IrAstOperation::Const(val) = const_init {
-                    val.as_constant(context)
+                    val.value.as_constant(context)
                 } else {
                     unreachable!("BUG! Initializer must be a const value.");
                 }
@@ -586,66 +663,90 @@ mod ir_builder {
         }));
 
         for block in fn_decl.blocks {
-            build_add_block_instructions(context, block, &named_blocks, &ptr_map, &mut arg_map);
+            build_add_block_instructions(
+                context,
+                block,
+                &named_blocks,
+                &ptr_map,
+                &mut arg_map,
+                md_map,
+                unresolved_calls,
+            );
         }
         Ok(())
     }
 
+    #[allow(clippy::type_complexity)]
     fn build_add_block_instructions(
         context: &mut Context,
         ir_block: IrAstBlock,
         named_blocks: &HashMap<String, Block>,
         ptr_map: &HashMap<String, Pointer>,
         val_map: &mut HashMap<String, Value>,
+        md_map: &HashMap<MdIdxRef, MetadataIndex>,
+        unresolved_calls: &mut Vec<(Block, Value, String, Vec<Value>, Option<MetadataIndex>)>,
     ) {
         let block = named_blocks.get(&ir_block.label).unwrap();
         for ins in ir_block.instructions {
+            let opt_ins_md_idx = ins.meta_idx.map(|mdi| md_map.get(&mdi).unwrap()).copied();
             let ins_val = match ins.op {
-                IrAstOperation::Asm(args, return_name, ops) => {
+                IrAstOperation::Asm(args, return_name, ops, meta_idx) => {
                     let args = args
                         .into_iter()
                         .map(|(name, opt_init)| AsmArg {
                             name,
                             initializer: opt_init.map(|init| match init {
                                 IrAstAsmArgInit::Var(var) => val_map.get(&var).cloned().unwrap(),
-                                IrAstAsmArgInit::Imm(cv) => cv.as_value(context),
+                                IrAstAsmArgInit::Imm(cv) => cv.value.as_value(
+                                    context,
+                                    md_map.get(cv.meta_idx.as_ref().unwrap()).copied(),
+                                ),
                             }),
                         })
                         .collect();
                     let body = ops
                         .into_iter()
-                        .map(|IrAstAsmOp { name, args, imm }| AsmInstruction {
-                            name,
-                            args,
-                            immediate: imm, //: Option<String>,
-                        })
+                        .map(
+                            |IrAstAsmOp {
+                                 name,
+                                 args,
+                                 imm,
+                                 meta_idx,
+                             }| AsmInstruction {
+                                name,
+                                args,
+                                immediate: imm,
+                                span_md_idx: meta_idx
+                                    .as_ref()
+                                    .and_then(|meta_idx| md_map.get(meta_idx).copied()),
+                            },
+                        )
                         .collect();
-                    block.ins(context).asm_block(args, body, return_name)
+                    let md_idx = meta_idx.map(|mdi| md_map.get(&mdi).unwrap()).copied();
+                    block
+                        .ins(context)
+                        .asm_block(args, body, return_name, md_idx)
                 }
                 IrAstOperation::Br(to_block_name) => {
                     let to_block = named_blocks.get(&to_block_name).unwrap();
-                    block.ins(context).branch(*to_block, None)
+                    block.ins(context).branch(*to_block, None, opt_ins_md_idx)
                 }
                 IrAstOperation::Call(callee, args) => {
-                    let function = context
-                        .functions
-                        .iter()
-                        .find_map(|(idx, content)| {
-                            if content.name == callee {
-                                Some(Function(idx))
-                            } else {
-                                None
-                            }
-                        })
-                        .unwrap();
-                    block.ins(context).call(
-                        function,
-                        &args
-                            .iter()
+                    // We can't resolve calls to other functions until we've done a first pass and
+                    // created them first.  So we can insert a NOP here, save the call params and
+                    // replace it with a CALL in a second pass.
+                    let nop = block.ins(context).nop();
+                    unresolved_calls.push((
+                        *block,
+                        nop,
+                        callee,
+                        args.iter()
                             .map(|arg_name| val_map.get(arg_name).unwrap())
                             .cloned()
                             .collect::<Vec<Value>>(),
-                    )
+                        opt_ins_md_idx,
+                    ));
+                    nop
                 }
                 IrAstOperation::Cbr(cond_val_name, true_block_name, false_block_name) => {
                     block.ins(context).conditional_branch(
@@ -653,26 +754,31 @@ mod ir_builder {
                         *named_blocks.get(&true_block_name).unwrap(),
                         *named_blocks.get(&false_block_name).unwrap(),
                         None,
+                        opt_ins_md_idx,
                     )
                 }
-                IrAstOperation::Const(val) => val.as_value(context),
+                IrAstOperation::Const(val) => val.value.as_value(context, opt_ins_md_idx),
                 IrAstOperation::ExtractElement(aval, ty, idx) => {
                     let ir_ty = ty.to_ir_aggregate_type(context);
                     block.ins(context).extract_element(
                         *val_map.get(&aval).unwrap(),
                         ir_ty,
                         *val_map.get(&idx).unwrap(),
+                        opt_ins_md_idx,
                     )
                 }
                 IrAstOperation::ExtractValue(val, ty, idcs) => {
                     let ir_ty = ty.to_ir_aggregate_type(context);
-                    block
-                        .ins(context)
-                        .extract_value(*val_map.get(&val).unwrap(), ir_ty, idcs)
+                    block.ins(context).extract_value(
+                        *val_map.get(&val).unwrap(),
+                        ir_ty,
+                        idcs,
+                        opt_ins_md_idx,
+                    )
                 }
-                IrAstOperation::GetPtr(src_name) => {
-                    block.ins(context).get_ptr(*ptr_map.get(&src_name).unwrap())
-                }
+                IrAstOperation::GetPtr(src_name) => block
+                    .ins(context)
+                    .get_ptr(*ptr_map.get(&src_name).unwrap(), opt_ins_md_idx),
                 IrAstOperation::InsertElement(aval, ty, val, idx) => {
                     let ir_ty = ty.to_ir_aggregate_type(context);
                     block.ins(context).insert_element(
@@ -680,6 +786,7 @@ mod ir_builder {
                         ir_ty,
                         *val_map.get(&val).unwrap(),
                         *val_map.get(&idx).unwrap(),
+                        opt_ins_md_idx,
                     )
                 }
                 IrAstOperation::InsertValue(aval, ty, ival, idcs) => {
@@ -689,11 +796,13 @@ mod ir_builder {
                         ir_ty,
                         *val_map.get(&ival).unwrap(),
                         idcs,
+                        opt_ins_md_idx,
                     )
                 }
-                IrAstOperation::Load(src_name) => {
-                    block.ins(context).load(*ptr_map.get(&src_name).unwrap())
-                }
+                IrAstOperation::Load(src_name) => block
+                    .ins(context)
+                    .load(*ptr_map.get(&src_name).unwrap(), opt_ins_md_idx),
+                IrAstOperation::Nop => block.ins(context).nop(),
                 IrAstOperation::Phi(pairs) => {
                     for (block_name, val_name) in pairs {
                         block.add_phi(
@@ -708,15 +817,82 @@ mod ir_builder {
                     let ty = ty.to_ir_type(context);
                     block
                         .ins(context)
-                        .ret(*val_map.get(&ret_val_name).unwrap(), ty)
+                        .ret(*val_map.get(&ret_val_name).unwrap(), ty, opt_ins_md_idx)
                 }
                 IrAstOperation::Store(stored_val_name, ptr_name) => block.ins(context).store(
                     *ptr_map.get(&ptr_name).unwrap(),
                     *val_map.get(&stored_val_name).unwrap(),
+                    opt_ins_md_idx,
                 ),
             };
             ins.value_name.map(|vn| val_map.insert(vn, ins_val));
         }
+    }
+
+    fn build_metadata_map(
+        context: &mut Context,
+        ir_metadata: &[(MdIdxRef, IrMetadatum)],
+    ) -> HashMap<MdIdxRef, MetadataIndex> {
+        let mut md_map = ir_metadata
+            .iter()
+            .filter_map(|(idx_ref, md)| match md {
+                IrMetadatum::FilePath(path) => Some((idx_ref, path)),
+                _otherwise => None,
+            })
+            .fold(HashMap::new(), |mut md_map, (idx_ref, path)| {
+                let path_content = Arc::from(std::fs::read_to_string(path).unwrap().as_str());
+                let md_idx = context.metadata.insert(Metadatum::FileLocation(
+                    Arc::new(path.clone()),
+                    path_content,
+                ));
+                md_map.insert(*idx_ref, MetadataIndex(md_idx));
+                md_map
+            });
+
+        for (idx_ref, md) in ir_metadata {
+            if let IrMetadatum::Span {
+                loc_idx,
+                start,
+                end,
+            } = md
+            {
+                let span_idx = context.metadata.insert(Metadatum::Span {
+                    loc_idx: md_map.get(loc_idx).copied().unwrap(),
+                    start: *start,
+                    end: *end,
+                });
+                md_map.insert(*idx_ref, MetadataIndex(span_idx));
+            }
+        }
+        md_map
+    }
+
+    #[allow(clippy::type_complexity)]
+    fn resolve_calls(
+        context: &mut Context,
+        unresolved_calls: Vec<(Block, Value, String, Vec<Value>, Option<MetadataIndex>)>,
+    ) -> Result<(), IrError> {
+        // All of the call instructions are currently NOPs which need to be replaced with actual
+        // calls.  We couldn't do it above until we'd gone and created all the functions first.
+        //
+        // Now we can loop and find the callee function for each call and replace the NOPs.
+        for (block, nop, callee, args, opt_ins_md_idx) in unresolved_calls {
+            let function = context
+                .functions
+                .iter()
+                .find_map(|(idx, content)| {
+                    if content.name == callee {
+                        Some(Function(idx))
+                    } else {
+                        None
+                    }
+                })
+                .unwrap();
+            let call_val =
+                Value::new_instruction(context, Instruction::Call(function, args), opt_ins_md_idx);
+            block.replace_instruction(context, nop, call_val)?;
+        }
+        Ok(())
     }
 }
 
