@@ -1,8 +1,9 @@
 pub const MANIFEST_FILE_NAME: &str = "Forc.toml";
+pub const LOCK_FILE_NAME: &str = "Forc.lock";
 pub const TEST_MANIFEST_FILE_NAME: &str = "Cargo.toml";
 pub const TEST_DIRECTORY: &str = "tests/";
 pub const SWAY_EXTENSION: &str = "sw";
-pub const FORC_DEPENDENCIES_DIRECTORY: &str = ".forc";
+pub const USER_FORC_DIRECTORY: &str = ".forc";
 pub const SRC_DIR: &str = "src";
 pub const SWAY_PREDICATE: &str = "predicate";
 pub const SWAY_LIBRARY: &str = "library";
