@@ -79,17 +79,26 @@ pub fn run(filter_regex: Option<regex::Regex>) {
         ("trait_import_with_star", ProgramState::Return(0)),
         ("tuple_desugaring", ProgramState::Return(9)),
         ("multi_item_import", ProgramState::Return(0)), // false
+        ("use_full_path_names", ProgramState::Return(1)),
         ("tuple_indexing", ProgramState::Return(1)),
         ("tuple_access", ProgramState::Return(42)),
         ("funcs_with_generic_types", ProgramState::Return(1)), // true
-        ("supertraits_1", ProgramState::Return(1)),
-        ("supertraits_2", ProgramState::Return(1)),
+        ("size_of", ProgramState::Return(1)),
+        ("supertraits", ProgramState::Return(1)),
+        ("new_allocator_test", ProgramState::Return(42)), // true
+        ("inline_if_expr_const", ProgramState::Return(0)),
+        ("method_on_empty_struct", ProgramState::Return(1)),
     ];
 
     let mut number_of_tests_run = positive_project_names.iter().fold(0, |acc, (name, res)| {
         if filter(name) {
             assert_eq!(crate::e2e_vm_tests::harness::runs_in_vm(name), *res);
-            assert_eq!(crate::e2e_vm_tests::harness::test_json_abi(name), Ok(()));
+            // cannot use partial eq on type `anyhow::Error` so I've used `matches!` here instead.
+            // https://users.rust-lang.org/t/issues-in-asserting-result/61198/3 for reference.
+            assert!(matches!(
+                crate::e2e_vm_tests::harness::test_json_abi(name),
+                Ok(())
+            ));
             acc + 1
         } else {
             acc
@@ -107,7 +116,7 @@ pub fn run(filter_regex: Option<regex::Regex>) {
         // when that is re-implemented we should reenable this test
         //"infinite_dependencies",
         "top_level_vars",
-        "dependencies_parsing_error",
+        "dependency_parsing_error",
         "disallowed_gm",
         "bad_generic_annotation",
         "bad_generic_var_annotation",
@@ -127,10 +136,9 @@ pub fn run(filter_regex: Option<regex::Regex>) {
         "star_import_alias",
         "item_used_without_import",
         "shadow_import",
-        "supertrait_dup_methods_1",
-        "supertrait_dup_methods_2",
-        "missing_supertrait",
         "missing_supertrait_impl",
+        "missing_func_from_supertrait_impl",
+        "supertrait_does_not_exist",
     ];
     number_of_tests_run += negative_project_names.iter().fold(0, |acc, name| {
         if filter(name) {
